+++ CHANGELOG.md
@@ -16,6 +16,7 @@ _No unreleased changes_
 ### Security 
 - Updated google-spreadsheet to 3.1.15 for CVE-2020-28168 (axios)
 - Updated electron to 10.2.0 for CVE-2020-26272
+- Updated ws to 7.4.6 for CVE-2021-32640
 
 ## [1.2.2] - 2020-09-18
 ### Security